[web] Move phase handling to the Main component
* It prevents re-rendering when the product is selected
  because it changes the product and the phase.
