+++ web/src/Main.test.jsx
@@ -21,33 +21,60 @@
 
 import React from "react";
 
-import { screen } from "@testing-library/react";
+import { act, screen } from "@testing-library/react";
 import { installerRender } from "./test-utils";
+import { createClient } from "./client";
+
+import { STARTUP, CONFIG, INSTALL } from "./client/phase";
+import { IDLE, BUSY } from "./client/status";
 
 import Main from "./Main";
 
 jest.mock("./client");
-jest.mock("./Questions", () => () => "Questions Mock");
-jest.mock("./Layout", () => ({ children }) => <>{children}</>);
+jest.mock("./Questions", () => () => <div>Questions Mock</div>);
+jest.mock("./DBusError", () => () => <div>D-BusError Mock</div>);
+jest.mock("./InstallationProgress", () => () => "InstallationProgress Mock");
+jest.mock("./InstallationFinished", () => () => "InstallationFinished Mock");
+jest.mock("./LoadingEnvironment", () => () => "LoadingEnvironment Mock");
 jest.mock('react-router-dom', () => ({
   Outlet: () => <div>Content</div>,
-  Navigate: () => <div>Navigate</div>,
 }));
 
-jest.mock("./context/software", () => ({
-  ...jest.requireActual("./context/software"),
-  useSoftware: () => {
+const callbacks = {};
+const getStatusFn = jest.fn();
+const getPhaseFn = jest.fn();
+
+// capture the latest subscription to the manager#onStatusChange for triggering it manually
+const onStatusChangeFn = cb => { callbacks.onStatusChange = cb };
+
+// capture the latest subscription to the manager#onPhaseChange for triggering it manually
+const onPhaseChangeFn = cb => { callbacks.onPhaseChange = cb };
+
+const changeStatusTo = status => act(() => callbacks.onStatusChange(status));
+const changePhaseTo = phase => act(() => callbacks.onPhaseChange(phase));
+
+beforeEach(() => {
+  createClient.mockImplementation(() => {
     return {
-      products: products,
-      selectedProduct: mockProduct
+      manager: {
+        getStatus: getStatusFn,
+        getPhase: getPhaseFn,
+        onPhaseChange: onPhaseChangeFn,
+        onStatusChange: onStatusChangeFn
+      },
+      monitor: {
+        onDisconnect: jest.fn()
+      }
     };
-  }
-}));
+  });
+});
 
-const products = [{ id: "Tumbleweed", name: "openSUSE Tumbleweed" }];
-let mockProduct = products[0];
+describe("when running on config phase", () => {
+  beforeEach(() => {
+    getPhaseFn.mockResolvedValue(CONFIG);
+    getStatusFn.mockResolvedValue(IDLE);
+  });
 
-describe("when a product is selected", () => {
   it("renders the Questions component and the content", async () => {
     installerRender(<Main />);
 
@@ -56,14 +83,59 @@ describe("when a product is selected", () => {
   });
 });
 
-describe("when no product is selected", () => {
+describe("when busy during startup", () => {
   beforeEach(() => {
-    mockProduct = null;
+    getPhaseFn.mockResolvedValue(STARTUP);
+    getStatusFn.mockResolvedValue(BUSY);
   });
 
-  it("redirects to the product selection page", async () => {
+  it("renders the Questions component and the content", async () => {
     installerRender(<Main />);
 
-    await screen.findByText("Navigate");
+    await screen.findByText("LoadingEnvironment Mock");
+  });
+});
+
+describe("when there are problems connecting with D-Bus service", () => {
+  beforeEach(() => {
+    getStatusFn.mockRejectedValue(new Error("Couldn't connect to D-Bus service"));
+  });
+
+  it("renders the DBusError component", async () => {
+    installerRender(<Main />);
+    await screen.findByText("D-BusError Mock");
+  });
+});
+
+describe("when D-Bus service status changes", () => {
+  beforeEach(() => {
+    getStatusFn.mockResolvedValue(BUSY);
+  });
+
+  it("renders InstallationProgress components when INSTALLING", async () => {
+    installerRender(<Main />);
+    await screen.findByText(/Loading.*environment/i);
+
+    changePhaseTo(INSTALL);
+    changeStatusTo(BUSY);
+    await screen.findByText("InstallationProgress Mock");
+  });
+
+  it("renders InstallationFinished components when INSTALLED", async () => {
+    installerRender(<Main />);
+    await screen.findByText(/Loading.*environment/i);
+
+    changePhaseTo(INSTALL);
+    changeStatusTo(IDLE);
+    await screen.findByText("InstallationFinished Mock");
+  });
+
+  it("renders the content if not PROBING, INSTALLING, or INSTALLED", async () => {
+    installerRender(<Main />);
+    await screen.findByText(/Loading.*environment/i);
+
+    changePhaseTo(CONFIG);
+    changeStatusTo(IDLE);
+    await screen.findByText("Content");
   });
 });
